+++ spec/rspec/its_spec.rb
@@ -69,6 +69,10 @@ module RSpec
             its("name") { is_expected.to eq("John") }
           end
 
+          context "using will_not" do
+            its("name") { will_not raise_error }
+          end
+
           context "using are_expected" do
             its("name.chars.to_a") { are_expected.to eq(%w[J o h n]) }
           end
@@ -103,6 +107,10 @@ module RSpec
                   should eq(64)
                 end.to raise_error(NoMethodError)
               end
+
+              context "using will" do
+                its(:age) { will raise_error(NoMethodError) }
+              end
             end
           end
 
@@ -117,6 +125,10 @@ module RSpec
               its([:not_here]) { should be_nil }
               its([:a, :ghost]) { should be_nil }
               its([:deep, :ghost]) { expect { should eq("missing") }.to raise_error(NoMethodError) }
+
+              context "using will" do
+                its([:deep, :ghost]) { will raise_error(NoMethodError) }
+              end
             end
           end
         end
@@ -129,6 +141,10 @@ module RSpec
               expect do
                 should eq("Symbol: a")
               end.to raise_error(NoMethodError)
+            end
+
+            context "using will" do
+              its([:a]) { will raise_error(NoMethodError) }
             end
           end
         end
@@ -238,6 +254,125 @@ module RSpec
             expect(example.metadata[:foo]).to be(true)
             expect(example.metadata[:bar]).to be(17)
           end
+        end
+      end
+
+      context "when expecting errors" do
+        subject do
+          Class.new do
+            def good; end
+
+            def bad
+              raise ArgumentError, "message"
+            end
+          end.new
+        end
+
+        its(:good) { will_not raise_error }
+        its(:bad) { will raise_error }
+        its(:bad) { will raise_error(ArgumentError) }
+        its(:bad) { will raise_error("message") }
+        its(:bad) { will raise_error(ArgumentError, "message") }
+      end
+
+      context "when expecting throws" do
+        subject do
+          Class.new do
+            def good; end
+
+            def bad
+              throw :abort, "message"
+            end
+          end.new
+        end
+
+        its(:good) { will_not throw_symbol }
+        its(:bad) { will throw_symbol }
+        its(:bad) { will throw_symbol(:abort) }
+        its(:bad) { will throw_symbol(:abort, "message") }
+      end
+
+      context "with change observation" do
+        subject do
+          Class.new do
+            attr_reader :count
+
+            def initialize
+              @count = 0
+            end
+
+            def increment
+              @count += 1
+            end
+
+            def noop; end
+          end.new
+        end
+
+        its(:increment) { will change { subject.count }.by(1) }
+        its(:increment) { will change { subject.count }.from(0) }
+        its(:increment) { will change { subject.count }.from(0).to(1) }
+        its(:increment) { will change { subject.count }.by_at_least(1) }
+        its(:increment) { will change { subject.count }.by_at_most(1) }
+
+        its(:noop) { will_not change { subject.count } }
+        its(:noop) { will_not change { subject.count }.from(0) }
+
+        its(:increment) do
+          expect { will_not change { subject.count }.by(0) }.to \
+            raise_error(NotImplementedError, '`expect { }.not_to change { }.by()` is not supported')
+        end
+
+        its(:increment) do
+          expect { will_not change { subject.count }.by_at_least(2) }.to \
+            raise_error(NotImplementedError, '`expect { }.not_to change { }.by_at_least()` is not supported')
+        end
+
+        its(:increment) do
+          expect { will_not change { subject.count }.by_at_most(3) }.to \
+            raise_error(NotImplementedError, '`expect { }.not_to change { }.by_at_most()` is not supported')
+        end
+      end
+
+      context "with output capture" do
+        subject do
+          Class.new do
+            def stdout
+              print "some output"
+            end
+
+            def stderr
+              $stderr.print "some error"
+            end
+
+            def noop; end
+          end.new
+        end
+
+        its(:stdout) { will output("some output").to_stdout }
+        its(:stderr) { will output("some error").to_stderr }
+
+        its(:noop) { will_not output("some error").to_stderr }
+        its(:noop) { will_not output("some output").to_stdout }
+      end
+
+      context "#will with non block expectations" do
+        subject do
+          Class.new do
+            def terminator
+              "back"
+            end
+          end.new
+        end
+
+        its(:terminator) do
+          expect { will be("back") }.to \
+            raise_error(ArgumentError, '`will` only supports block expectations')
+        end
+
+        its(:terminator) do
+          expect { will_not be("back") }.to \
+            raise_error(ArgumentError, '`will_not` only supports block expectations')
         end
       end
     end